Create README.md
This is a two player game in which you have to control a circle on the screen(evil circle) that catches the bouncing balls on the screen. Two players can play the game and for it the controls are as follows Player 1 can go in 4 directions with W,A,S and D keys. Player 2 can do the same with arrow UP, DOWN, RIGHT and LEFT.